requirements.txt
# --- Servidor y Protocolo (CRÍTICO para HTTP/2) ---
fastapi
uvicorn[standard]
h2
gunicorn

# --- LangChain stack ---
langchain
langchain-core
langchain-experimental
langchain-google-genai

# --- Google Cloud ---
google-cloud-storage

# --- API y Comunicaciones ---
python-multipart>=0.0.9
httpx

# --- NLP / ML / Datos ---
transformers
numpy
pandas
openpyxl

# --- Validaciones y Utils ---
pydantic
pydantic[email]
python-dotenv
pyyaml
Torch
# --- Base de Datos ---
psycopg2-binary

# --- Manejo de Archivos ---
pymupdf
h2
https://carlos.alberto.diaz@segurosbolivar.com:cmVmdGtuOjAxOjE3NzM1MTcyOTM6d0dLcmdIOEFaY0F2MHpZNEJCTGFZZkdRUUJ5@segurosbolivar.jfrog.io/artifactory/transversal-ia-langchain-python-lib-dev-virtual/ia_transversal_langchain_python_lib/2.6.3/ia_transversal_langchain_python_lib-2.6.3-py3-none-any.whl
